Drop SOC 2 template and the catalog; group policy types under policy/
**SOC 2 removed.** It was the one shipped posture with no
data-category regulation behind it — SOC 2's Common Criteria
enumerate no identifier list, so the label scope was industry
secrets-scanning practice wearing a framework's name. Its own
docstring said as much while the module header and the user-visible
description claimed to "cover" CC6.1 / CC6.7 and to "satisfy" ISO
27001 A.9.2.4 (a 2013-edition control, superseded by A.5.17 in 2022,
and a secret-management process no redaction pass can satisfy). A
secrets scan is a good feature; it is not a compliance template.

**`TemplateCatalog` removed.** A `(id, version)`-keyed registry with
serde shims, version lookup, and collision tests is a lookup
structure, not a regulatory posture. Callers who want one build it
over the `PolicyTemplate` variants; the crate ships the postures.
This also retires the `try_from` deserialization and shipped-pairing
tests added earlier in this branch, which existed only to make the
catalog's own invariants hold, along with the integration smoke test
that iterated it.

**Governance grouped under `policy/`.** `label`, `predicate`, and
`rule` join `PolicyDefinition` in a `policy` module; `redaction`
stays at the crate root, since operator vocabulary is what rules
dispatch *to* rather than part of the policy structure itself.
`lib.rs` is now crate docs plus re-exports. The public API is
unchanged except `Predicate`, which moves from
`elide_governance::predicate::Predicate` to `elide_governance::Predicate`.

Also corrects the crate-level docs, which still described the engine
stamping `group:<policy_id>:<name>` tags at request-compile time —
the same stale claim already fixed in `label.rs` and `predicate.rs`.

Verified: clippy clean, docs build without warnings, 65 tests pass.
