+++ crates/elide-pipeline/tests/policy_shapes.rs
@@ -8,7 +8,7 @@
 use bytes::Bytes;
 use elide_core::entity::LabelRef;
 use elide_core::entity::audit::AuditKind;
-use elide_governance::predicate::Predicate;
+use elide_governance::Predicate;
 use elide_governance::redaction::{ModalityRedactions, TextRedaction};
 use elide_governance::{
     LabelEntry, LabelGroup, Labels, PolicyDefinition, PolicyRule, RuleDispatch,
