+++ crates/elide-template/src/template/ccpa.rs
@@ -24,7 +24,7 @@
 //! [`Pseudonymize`]: elide_governance::redaction::TextRedaction::Pseudonymize
 
 use elide_core::entity::LabelRef;
-use elide_governance::predicate::Predicate;
+use elide_governance::Predicate;
 use elide_governance::redaction::{ModalityRedactions, TextRedaction};
 use elide_governance::{LabelGroup, Labels, PolicyDefinition, PolicyRule, RuleDispatch};
 use jiff::civil::Date;
