+++ crates/elide-template/src/template/gdpr/erase.rs
@@ -1,4 +1,4 @@
-use elide_governance::predicate::Predicate;
+use elide_governance::Predicate;
 use elide_governance::redaction::{ModalityRedactions, TextRedaction};
 use elide_governance::{LabelGroup, Labels, PolicyDefinition, PolicyRule, RuleDispatch};
 use semver::Version;
