+++ crates/elide-template/src/template/mod.rs
@@ -7,7 +7,6 @@ pub(crate) mod ccpa;
 pub(crate) mod gdpr;
 pub(crate) mod hipaa;
 pub(crate) mod pci;
-pub(crate) mod soc2;
 
 use elide_governance::PolicyDefinition;
 use hipstr::HipStr;
